requirements.txt
ipykernel==6.28.0
fastapi==0.109.0
uvicorn==0.25.0
langchain-anthropic==0.3.1
langchain-core==0.3.29
langchain==0.3.14
pydantic==2.10.4
langchain-community==0.3.14
requests==2.32.3
rank-bm25
streamlit
boto3
botocore
numpy
langchain-huggingface
sentence-transformers
torch
faiss-cpu
langchain-groq
pypdf